ACCNT.03 - Ability to add funds to account

+++ IntegralInvesting/DataAccessWebAPI/Models/BankAccount.cs
@@ -13,7 +13,7 @@ namespace DataAccessWebAPI.Models
         public string UserId { get; set; }
 
         [Column(TypeName = "decimal(8, 2)")]
-        public float CurrentFunds { get; set; }
+        public decimal CurrentFunds { get; set; }
 
         [Required]
         [MaxLength(60)]
